Add Refinance modal with mock data
