+++ src/pages/LoansPage/components/LoansActiveTable/LoansActiveTable.module.less
@@ -162,32 +162,6 @@
   }
 }
 
-.actionsButtons {
-  display: flex;
-  align-items: center;
-  justify-content: flex-end;
-  gap: 8px;
-  width: 100%;
-}
-
-.refinanceBtn {
-  flex-grow: 1;
-}
-
-.additionalRepayInfo {
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-  padding: 8px 0 8px;
-}
-.mainRepayInfo {
-  display: flex;
-  justify-content: flex-start;
-}
-.repayButton {
-  width: 100%;
-}
-
 .sortDropdown {
   & > button {
     min-width: 230px;
