+++ src/pages/LoansPage/components/LoansActiveTable/TableCells/ActionsCell/ActionsCell.module.less
@@ -1,0 +1,122 @@
+.actionsButtons {
+  display: flex;
+  align-items: center;
+  justify-content: flex-end;
+  gap: 8px;
+  width: 100%;
+}
+
+.refinanceBtn {
+  flex-grow: 1;
+}
+
+.repayModalInfo {
+  display: flex;
+  justify-content: flex-start;
+}
+
+.repayModalAdditionalInfo {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+  padding: 8px 0 8px;
+}
+
+.repayModalButton,
+.refinanceModalButton {
+  width: 100%;
+}
+
+.currentLoanInfo {
+  margin-bottom: 18px;
+}
+
+.newLoanInfo {
+  margin-bottom: 28px;
+}
+
+.difference {
+  margin-bottom: 16px;
+}
+
+//? LoanInfo component
+.loanInfo {
+  display: flex;
+  flex-direction: column;
+  align-items: stretch;
+  gap: 8px;
+  color: var(--content-primary);
+}
+.loanInfoFaded {
+  color: var(--content-secondary);
+}
+
+.loanInfoTitle {
+  font: var(--title-sm);
+  text-transform: uppercase;
+  text-align: center;
+
+  @media (max-width: 640px) {
+    font: var(--title-xs);
+  }
+}
+
+.loanInfoStats {
+  display: flex;
+  justify-content: space-around;
+}
+
+.loanInfoValue {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+
+  p:first-child {
+    font: var(--important-text-xl);
+
+    @media (max-width: 640px) {
+      font: var(--important-text-md);
+    }
+  }
+
+  p:last-child {
+    font: var(--table-header-md);
+    text-transform: uppercase;
+
+    @media (max-width: 640px) {
+      font: var(--table-header-sm);
+    }
+  }
+}
+
+//? LoanDifference component
+.loanDifference {
+  border: var(--border-less-primary);
+  border-style: dashed;
+  padding: 16px;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.loanDifferenceTitle {
+  font: var(--important-text-xl);
+  color: var(--additional-green-primary-deep);
+
+  @media (max-width: 640px) {
+    font: var(--important-text-md);
+  }
+}
+.loanDifferenceTitleRed {
+  color: var(--additional-red-primary-deep);
+}
+
+.loanDifferenceSubtitle {
+  font: var(--table-header-md);
+  text-transform: uppercase;
+
+  @media (max-width: 640px) {
+    font: var(--table-header-sm);
+  }
+}
