+++ src/pages/LoansPage/components/LoansActiveTable/TableCells/ActionsCell/index.ts
@@ -1,0 +1,1 @@
+export * from './ActionsCell'
