+++ src/pages/LoansPage/components/LoansActiveTable/TableCells/index.ts
@@ -1,4 +1,4 @@
-export * from './ActionsCell'
+export * from './ActionsCell/ActionsCell'
 export * from './DebtCell'
 export * from './StatusCell'
 export * from './HealthCell'
